requirements.txt
matplotlib
pandas
numpy
streamlit
plotly
google-cloud-bigquery
db-dtypes